backend/requirements.txt
# -------------------------
# Core framework
# -------------------------
Django>=5.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3

# -------------------------
# CORS (frontend integration)
# -------------------------
django-cors-headers>=4.3

# -------------------------
# API & utilities
# -------------------------
python-dateutil>=2.9

# -------------------------
# Testing
# -------------------------
pytest>=8.0
pytest-django>=4.8
behave>=1.2
behave-django>=1.4

# -------------------------
# Code quality (used in CI)
# -------------------------
ruff>=0.6
black>=24.4

# -------------------------
# AI Integration
# -------------------------
openai>=1.0.0

# -------------------------
# Document ingestion (PDF + OCR)
# -------------------------
pymupdf>=1.24

# -------------------------
# KSeF (Polish e-invoicing) — requires Python 3.12+
# -------------------------
ksef2>=0.1

# -------------------------
# Payments (portal deposits)
# -------------------------
stripe>=11.0,<14.0

# -------------------------
# Environment variables
# -------------------------
python-dotenv>=1.0.0

# -------------------------
# Production / AWS
# -------------------------
gunicorn>=21.0
whitenoise>=6.6
psycopg2-binary>=2.9
boto3>=1.35
